+++ src/main/java/ru/yandex/money/common/dbqueue/settings/TaskRetryType.java
@@ -9,13 +9,15 @@ package ru.yandex.money.common.dbqueue.settings;
 public enum TaskRetryType {
 
     /**
-     * Задача откладывается по геометрической прогрессии в минутах.
+     * Задача откладывается по геометрической прогрессии относительно интервала
+     * {@link QueueSettings#getRetryInterval()}
      * Знаменатель прогрессии равен двум.
      * Первые 6 членов: 1 2 4 8 16 32
      */
     GEOMETRIC_BACKOFF,
     /**
-     * Задача откладывается по арифметической прогрессии в минутах.
+     * Задача откладывается по арифметической прогрессии относительно интервала
+     * {@link QueueSettings#getRetryInterval()}.
      * Разность прогрессии равна двум.
      * Первые 6 членов: 1 3 5 7 9 11
      */
@@ -24,7 +26,7 @@ public enum TaskRetryType {
      * Задача откладывает с фиксированной задержкой.
      * <p>
      * Значение фиксированной задержки устанавливается через настройку
-     * {@link QueueSettings.AdditionalSetting#RETRY_FIXED_INTERVAL_DELAY}
+     * {@link QueueSettings#getRetryInterval()}
      */
-    FIXED_INTERVAL
+    LINEAR_BACKOFF
 }