+++ src/test/java/ru/yandex/money/common/dbqueue/dao/QueueDaoTest.java
@@ -85,7 +85,7 @@ public class QueueDaoTest extends BaseDaoTest {
 
         ZonedDateTime beforeExecution = ZonedDateTime.now();
         Duration executionDelay = Duration.ofHours(1L);
-        Boolean reenqueueResult = executeInTransaction(() -> queueDao.reenqueue(location, enqueueId, executionDelay, true));
+        Boolean reenqueueResult = executeInTransaction(() -> queueDao.reenqueue(location, enqueueId, executionDelay));
         Assert.assertThat(reenqueueResult, equalTo(true));
         jdbcTemplate.query("select * from " + QueueDatabaseInitializer.DEFAULT_TABLE_NAME + " where id=" + enqueueId, rs -> {
             ZonedDateTime afterExecution = ZonedDateTime.now();
@@ -115,7 +115,7 @@ public class QueueDaoTest extends BaseDaoTest {
         });
 
         Boolean reenqueueResult = executeInTransaction(() ->
-                queueDao.reenqueue(location, enqueueId, Duration.ofHours(1L), true));
+                queueDao.reenqueue(location, enqueueId, Duration.ofHours(1L)));
 
         Assert.assertThat(reenqueueResult, equalTo(true));
         jdbcTemplate.query("select * from " + QueueDatabaseInitializer.DEFAULT_TABLE_NAME + " where id=" + enqueueId, rs -> {
@@ -126,37 +126,10 @@ public class QueueDaoTest extends BaseDaoTest {
     }
 
     @Test
-    public void reenqueue_should_not_reset_attempts() throws Exception {
-        QueueLocation location = generateUniqueLocation();
-        Long enqueueId = executeInTransaction(() ->
-                queueDao.enqueue(location, new EnqueueParams<>()));
-        executeInTransaction(() -> {
-            jdbcTemplate.update("update " + QueueDatabaseInitializer.DEFAULT_TABLE_NAME + " set attempt=10 where id=" + enqueueId);
-        });
-
-        jdbcTemplate.query("select * from " + QueueDatabaseInitializer.DEFAULT_TABLE_NAME + " where id=" + enqueueId, rs -> {
-            Assert.assertThat(rs.next(), equalTo(true));
-            Assert.assertThat(rs.getLong("attempt"), equalTo(10L));
-            return new Object();
-        });
-
-        Boolean reenqueueResult = executeInTransaction(() ->
-                queueDao.reenqueue(location, enqueueId, Duration.ofHours(1L), false));
-
-        Assert.assertThat(reenqueueResult, equalTo(true));
-        jdbcTemplate.query("select * from " + QueueDatabaseInitializer.DEFAULT_TABLE_NAME + " where id=" + enqueueId, rs -> {
-            Assert.assertThat(rs.next(), equalTo(true));
-            Assert.assertThat(rs.getLong("attempt"), equalTo(10L));
-            return new Object();
-        });
-    }
-
-
-    @Test
     public void reenqueue_should_return_false_when_no_update() throws Exception {
         QueueLocation location = generateUniqueLocation();
         Boolean reenqueueResult = executeInTransaction(() ->
-                queueDao.reenqueue(location, 0L, Duration.ofHours(1L), true));
+                queueDao.reenqueue(location, 0L, Duration.ofHours(1L)));
         Assert.assertThat(reenqueueResult, equalTo(false));
     }